[REFACTOR] Changed the folder structure and added parallelism to multiplication
